requirements.txt
# Bot & async
aiogram>=3.15.0
aiohttp>=3.11.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0

# Validation & config
pydantic>=2.10.0
pydantic-settings>=2.6.0
typing-extensions>=4.12.0

# Task queue
celery[redis]>=5.4.0
redis>=5.2.0

# HTTP client
httpx>=0.28.0

# PDF & OCR (LLM via OpenRouter)
pypdf>=5.0.0
pdf2image>=1.17.0
Pillow>=11.0.0
openai>=1.55.0

# Utils
python-dotenv>=1.0.0
openpyxl>=3.1.0
sentry-sdk>=2.0.0
python-json-logger>=2.0.7

# Tests
pytest>=8.3.0
pytest-asyncio>=0.24.0
